refactor: 각 Manager 별로 OutputView 분리
