+++ src/main/java/subway/LineManager.java
@@ -1,7 +1,7 @@
 package subway;
 
 import java.util.Scanner;
-import view.OutputView;
+import view.LineManagerOutputView;
 
 public class LineManager {
 
@@ -26,7 +26,7 @@ public class LineManager {
     public static String userOption = "";
 
     public static void manage(Scanner scanner) {
-        OutputView.printLineManagerMainScreen();
+        LineManagerOutputView.printLineManagerMainScreen();
         userOption = getUserOption(scanner);
         callOptionMenu(userOption, scanner);
     }
