+++ src/main/java/subway/StationManager.java
@@ -2,12 +2,11 @@ package subway;
 
 import java.util.Scanner;
 import subway.domain.Station;
-import view.OutputView;
+import view.StationManagerOutputView;
 
 public class StationManager {
 
     public static final String WANT_QUIT_CODE = "B";
-    public static final String QUIT_OPTION = "B. 돌아가기";
     public static final String ASK_OPTION_MESSAGE = "## 원하는 기능을 선택하세요.";
     public static final String ERROR_PREFIX = "[ERROR] ";
     public static final String OPTION_ERROR_MESSAGE = "1~3 또는 B 옵션 중 하나를 입력하세요";
@@ -26,7 +25,7 @@ public class StationManager {
     public static String userOption = "";
 
     public static void manage(Scanner scanner) {
-        OutputView.printStationManagerMainScreen();
+        StationManagerOutputView.printStationManagerMainScreen();
         userOption = getUserOption(scanner);
         callOptionMenu(userOption, scanner);
     }
